+++ shared/src/commonMain/kotlin/kfu/itis/freshnews/feature/auth/domain/AuthRepository.kt
@@ -7,8 +7,8 @@ interface AuthRepository {
 
     suspend fun signUp(login: String, password: String): Flow<UserProfile?>
     fun signIn(login: String, password: String): Flow<UserProfile?>
-    suspend fun saveUserId(id: Int)
-    suspend fun getUserId(): Int?
+    suspend fun saveUserId(id: Long)
+    suspend fun getUserId(): Long?
     suspend fun setWelcomeScreenShown()
     suspend fun isWelcomeScreenShown(): Boolean
 }
